feat(2023): Day 10: Pipe Maze

+++ 2023/day10/part1.py
@@ -3,4 +3,4 @@ from helpers import find_longest_cycle
 
 def part1(puzzle_input):
     (cycle, _) = find_longest_cycle(puzzle_input)
-    return len(cycle) / 2
+    return int(len(cycle) / 2)
